Running Version with posts

+++ database/migrations/2020_05_22_124010_create_post_tag_table.php
@@ -16,12 +16,12 @@ class CreatePostTagTable extends Migration
         Schema::create('post_tag', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedInteger('post_id')->index();
+            $table->unsignedBigInteger('post_id')->index();
             $table->foreign('post_id')->references('id')
                 ->on('posts')
                 ->onDelete('cascade');
 
-            $table->unsignedInteger('tag_id')->index();
+            $table->unsignedBigInteger('tag_id')->index();
             $table->foreign('tag_id')->references('id')
                 ->on('tags')
                 ->onDelete('cascade');
